refactor jobs and processors
